Adds ability to make sure duplicates have same time
If two "duplicate images" don't have the same time, then they are
considered different images. I would consider this the first line of
defense when looking for duplicates. These are the easy ones to delete.
Once those are taken care of, it is good to turn this flag off and run
it again.